Audio and Video working at the same time much better. RtspServer now sends correct Setup and Play responses. (QuickTime works again and seq= and rtptime= are added to Rtp-Info header as per RFC2326
Removed redundant calls to recieve in RtpClient. Changed the order in recieving to allow Rtcp first before Rtp.

Fixed bugs with RtcpPacket parsing and made more efficient (Length on SourceDescriptionItem and a few others)

Added methods to MediaDescription for adding and removing lines internally.

Updated comments and code to support examples.

+++ Rtcp/ReceiversReport.cs
@@ -24,7 +24,7 @@ namespace Media.Rtcp
 
         public ReceiversReport(byte[] packet, int offset) 
         {
-            SynchronizationSourceIdentifier = (uint)System.Net.IPAddress.NetworkToHostOrder((int)BitConverter.ToInt32(packet, offset + 0));
+            SynchronizationSourceIdentifier = (uint)System.Net.IPAddress.NetworkToHostOrder(BitConverter.ToInt32(packet, offset + 0));
             offset += 4; 
             if (packet.Length > 4)
             {
